[TS-346] Refactor: addStar 함수를 userThunk로 이동
- API 문서 분류에 맞게 addStar 함수를 user 쪽으로 이동
