+++ src/api/user/userThunk.ts
@@ -147,3 +147,12 @@ export const selectStar = createAsyncThunk(
 		}
 	},
 );
+
+export const addStar = createAsyncThunk("star/addStar", async (_, thunkOptions) => {
+	try {
+		const { data } = await axiosInstance.put(END_POINTS.ADD_STAR);
+		return data;
+	} catch (error) {
+		throw thunkOptions.rejectWithValue(error);
+	}
+});
